+++ v3/errors.go
@@ -45,6 +45,15 @@ func (e ErrEmptyKey) Error() string {
 	return "zero-length key, or key only contains whitespace"
 }
 
+// ErrEmptyEnvStack is returned whenever you call a method on an empty EnvStack
+type ErrEmptyEnvStack struct {
+	method string
+}
+
+func (e ErrEmptyEnvStack) Error() string {
+	return fmt.Sprintf("env stack is empty; %s", e.method)
+}
+
 // ErrNilPointer is returned whenever you call a method on the Env struct
 // with a nil pointer
 type ErrNilPointer struct {